Move epochs to libsyntax
