+++ src/libsyntax/lib.rs
@@ -23,6 +23,7 @@
 #![feature(unicode)]
 #![feature(rustc_diagnostic_macros)]
 #![feature(match_default_bindings)]
+#![feature(non_exhaustive)]
 #![feature(i128_type)]
 #![feature(const_atomic_usize_new)]
 #![feature(rustc_attrs)]
@@ -114,6 +115,7 @@ pub mod codemap;
 #[macro_use]
 pub mod config;
 pub mod entry;
+pub mod epoch;
 pub mod feature_gate;
 pub mod fold;
 pub mod parse;
